+++ recipes/pagerduty.rb
@@ -27,7 +27,7 @@ end
 if using_old_pagerduty_key_attribute?
   Chef::Log.warn('The nagios.pagerduty_key attribute is deprecated. It is replaced by the nagios.pagerduty.key attribute.')
   Chef::Log.warn('Assigning nagios.pagerduty.key from nagios.pagerduty_key now.')
-  node.normal['nagios']['pagerduty']['key'] = node['nagios']['pagerduty_key']
+  node.default['nagios']['pagerduty']['key'] = node['nagios']['pagerduty_key']
 end
 
 package 'perl-CGI' do